+++ app/src/main/java/com/example/anywherefitness/Api/UserCrudApi.kt
@@ -2,17 +2,12 @@ package com.example.anywherefitness.Api
 
 import com.example.anywherefitness.model.CreateUser
 import com.example.anywherefitness.model.FitnessClass
-import com.example.anywherefitness.model.FitnessClassResult
 import com.example.anywherefitness.model.User
 import com.example.anywherefitness.ui.Instructor.CreateFitnessClass
 import retrofit2.Call
 import retrofit2.http.*
 
-//User
-//username, password, id, isInstructor
 interface UserCrudApi {
-    @GET("users")
-    fun getUsers(): Call<List<User>>
 
     @GET("users/{id}")
     fun getUserById(@Path("id") id: Int): Call<User>
@@ -26,6 +21,7 @@ interface UserCrudApi {
     @GET("classes")
     fun getAllClasses(@Header ("Authorization") token: String): Call<List<FitnessClass>>
 
+    //These functions are unused because the API does not support clients registering for classes
     @POST("classes/join/{id}")
     fun addUserToClass(@Path("id") id: Int,
                        @Body userId: Int): Call<FitnessClass>
