+++ app/src/main/java/com/example/anywherefitness/ui/Instructor/CreateClassViewModel.kt
@@ -1,30 +1,30 @@
 package com.example.anywherefitness.ui.Instructor
 
 import android.app.Application
-import android.view.View
-import android.widget.Button
+import android.util.Log
 import android.widget.EditText
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
+import com.example.anywherefitness.Api.UserApiBuilder
 import com.example.anywherefitness.App
-import com.example.anywherefitness.database.UserRepo
 import com.example.anywherefitness.model.FitnessClass
 import com.example.anywherefitness.model.User
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
 class CreateClassViewModel(application: Application) : AndroidViewModel(application) {
 
-    private var repo: UserRepo = UserRepo(application)
+    //private var repo: UserRepo = UserRepo(application)
+
+    var responseString = MutableLiveData<String>()
 
     private val _text = MutableLiveData<String>().apply {
         value = "Create Class"
     }
     val text: LiveData<String> = _text
 
-    //TODO: add sync with api when new class is created, can likely just do it when create class button is clicked because it won't be done
-    //as frequently
-    //TODO: fix things after changing class constructor to match api
     fun createClass(etName: EditText,
                     etType: EditText,
                     etTime: EditText,
@@ -34,11 +34,8 @@ class CreateClassViewModel(application: Application) : AndroidViewModel(applicat
                     etAttendees: EditText,
                     etSize: EditText): FitnessClass {
 
-        //can't convert et.text to Int in constructor
         val duration = etDuration.text.toString()
-        //val durationInt = duration.toInt()
         val intensity = etIntensity.text.toString()
-        //val intensityInt = intensity.toInt()
         val attendees = etAttendees.text.toString()
         val attendeesInt = attendees.toInt()
         val size = etSize.text.toString()
@@ -57,13 +54,27 @@ class CreateClassViewModel(application: Application) : AndroidViewModel(applicat
         return newClass
     }
 
-    fun insert(newClass: FitnessClass) {
-        repo.insert(newClass)
+    fun addClassToApi(getSavedToken: String?, newCreateClass: CreateFitnessClass){
+
+        UserApiBuilder.userRetro().addClass(getSavedToken!!, newCreateClass).enqueue(object :
+            Callback<Void> {
+            override fun onFailure(call: Call<Void>, t: Throwable) {
+                Log.i("BIGBRAIN", "onFailure $t")
+                responseString.value = "Failed to create class ${newCreateClass.name}"
+            }
+
+            override fun onResponse(call: Call<Void>, response: Response<Void>) {
+                if (response.isSuccessful) {
+                    Log.i("BIGBRAIN", "success ${response.body()}")
+                    responseString.value = "Successfully created new class ${newCreateClass.name}"
+                } else {
+                    Log.i("BIGBRAIN", "onFailure $response")
+                    responseString.value = "Failed to create class ${newCreateClass.name}"
+                }
+            }
+        })
     }
 
-    fun saveClass(newClass: FitnessClass) {
-        //make appropriate db calls and ui updates here
-    }
 
     fun getUser(): LiveData<User>?{
         return App.repo?.getUser()
