+++ app/src/main/java/com/example/anywherefitness/ui/Instructor/CreateClassesFragment.kt
@@ -2,23 +2,17 @@ package com.example.anywherefitness.ui.Instructor
 
 import android.content.Context
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
-import com.example.anywherefitness.Api.UserApiBuilder
 import com.example.anywherefitness.R
 import com.example.anywherefitness.model.User
 import com.example.anywherefitness.ui.LoginActivity
 import kotlinx.android.synthetic.main.fragment_create_class_instructor.*
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 class CreateClassesFragment : Fragment() {
 
@@ -53,34 +47,23 @@ class CreateClassesFragment : Fragment() {
                 et_class_attendees,
                 et_class_size)
 
-            //TODO: put this in view model
-            //TODO: figure out how to get class_Id back
-            //TODO: get instructor_id programmatically
-            println(userId)
-            //will either need to update in the my classes fragment, or figure out something here
-            val newCreateClass = CreateFitnessClass(newClass.name!!, newClass.type!!, newClass.location!!, newClass.duration!!, newClass.intensity!!,
-                newClass.max_size.toString(), newClass.starttime, userId, null)
+            val newCreateClass = CreateFitnessClass(newClass.name ?: "",
+                newClass.type ?: "",
+                newClass.location ?: "",
+                newClass.duration ?: "",
+                newClass.intensity ?: "",
+                newClass.max_size.toString(),
+                newClass.starttime,
+                userId,
+                null)
 
             val saveToken = activity?.getSharedPreferences(LoginActivity.SAVE_TOKEN, Context.MODE_PRIVATE)
             val getSavedToken = saveToken?.getString(LoginActivity.GET_SAVE_TOKEN, "default")
 
-            UserApiBuilder.userRetro().addClass(getSavedToken!!, newCreateClass).enqueue(object: Callback<Void> {
-                override fun onFailure(call: Call<Void>, t: Throwable) {
-                    Toast.makeText(context, "onFailure Failed to create class", Toast.LENGTH_LONG).show()
-                    Log.i("BIGBRAIN", "onFailure $t")
-                }
+            createClassViewModel.addClassToApi(getSavedToken, newCreateClass)
 
-                override fun onResponse(call: Call<Void>, response: Response<Void>) {
-                    if (response.isSuccessful) {
-                        //createClassViewModel.insert(newClass)
-                        Toast.makeText(context, "new class created ${newClass.name}", Toast.LENGTH_LONG).show()
-                        Log.i("BIGBRAIN", "success ${response.body()}")
-                    } else {
-                        Toast.makeText(context, "onResponse failed to create class", Toast.LENGTH_LONG).show()
-                        Log.i("BIGBRAIN", "onFailure $response")
-                    }
-                }
-            })
+            val responseObserver = Observer<String> { Toast.makeText(context, it, Toast.LENGTH_LONG).show() }
+            createClassViewModel.responseString.observe(this, responseObserver)
         }
     }
 }
