+++ app/src/main/java/com/example/anywherefitness/ui/Instructor/InstructorActivity.kt
@@ -9,18 +9,13 @@ import com.google.android.material.bottomnavigation.BottomNavigationView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.edit
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProviders
 import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
 import com.example.anywherefitness.App
 import com.example.anywherefitness.R
-import com.example.anywherefitness.model.User
 import com.example.anywherefitness.ui.LoginActivity
-import com.example.anywherefitness.ui.StartUpActivity
-import com.example.anywherefitness.ui.client.MyClassesFragment
-import com.example.anywherefitness.viewmodel.LoginViewModel
 
 class InstructorActivity : AppCompatActivity() {
