+++ app/src/main/java/com/example/anywherefitness/ui/client/FindClassesFragment.kt
@@ -45,24 +45,17 @@ class FindClassesFragment : Fragment() {
         val getSavedToken = saveToken?.getString(LoginActivity.GET_SAVE_TOKEN, "default")
         findClassesViewModel.getList(getSavedToken!!)
 
-
         val list = mutableListOf<FitnessClass>()
-
-
-
 
         rv_find_classes.layoutManager = LinearLayoutManager(context)
         rv_find_classes.adapter = SearchFitnessClassAdapter(list)
 
         val something = Observer<List<FitnessClass>> {
             list.addAll(it)
-            list.add(it[it.size - 1])
             rv_find_classes.adapter?.notifyDataSetChanged()
         }
         
         findClassesViewModel.resultList.observe(this, something)
-
-        //something.onChanged(findClassesViewModel.resultList.value)
 
         sv_find.setOnQueryTextListener(object : SearchView.OnQueryTextListener{
             val newList = mutableListOf<FitnessClass>()
@@ -89,16 +82,9 @@ class FindClassesFragment : Fragment() {
                 if (p0.isNullOrBlank()){
                     rv_find_classes.adapter?.notifyDataSetChanged()
                 }
-
-
                 return true
             }
-
         })
-        
-
-
-
     }
 
     inner class SearchFitnessClassAdapter (val fitnessClassList: MutableList<FitnessClass>):
@@ -141,7 +127,6 @@ class FindClassesFragment : Fragment() {
             }
         }
 
-
         inner class ViewHolder(view: View): RecyclerView.ViewHolder(view) {
             val fitnessClassName: TextView = view.tv_class_name
             val fitnessClassType: TextView = view.tv_class_type
@@ -155,8 +140,4 @@ class FindClassesFragment : Fragment() {
             val fitnessClassId: TextView = view.tv_class_id
         }
     }
-
-
-
-
 }
