+++ app/src/main/java/com/example/anywherefitness/viewmodel/LoginModelView.kt
@@ -1,18 +1,11 @@
 package com.example.anywherefitness.viewmodel
 
-import android.app.Activity
-import android.content.Context
-import android.content.Intent
-import androidx.core.content.ContextCompat.startActivity
-import androidx.core.content.edit
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.anywherefitness.Api.UserApiBuilder
 import com.example.anywherefitness.App
 import com.example.anywherefitness.model.User
-import com.example.anywherefitness.ui.LoginActivity
-import com.example.anywherefitness.ui.StartUpActivity
 import org.json.JSONObject
 import retrofit2.Call
 import retrofit2.Callback
@@ -22,8 +15,6 @@ class LoginViewModel : ViewModel(){
 
     val userToken = MutableLiveData<String>()
     val currentUser = MutableLiveData<User>()
-
-
 
     fun getToken(user: User) {
 
@@ -39,7 +30,6 @@ class LoginViewModel : ViewModel(){
                     userToken.value = null
                 }
             }
-
         })
     }
 
@@ -64,7 +54,6 @@ class LoginViewModel : ViewModel(){
                     currentUser.value = null
                 }
             }
-
         })
     }
 
@@ -75,9 +64,6 @@ class LoginViewModel : ViewModel(){
     fun getUser(): LiveData<User>?{
         return App.repo?.getUser()
     }
-
-
-
 }
