+++ app/src/main/java/com/example/anywherefitness/viewmodel/RegisterViewModel.kt
@@ -1,14 +1,9 @@
 package com.example.anywherefitness.viewmodel
 
-import android.content.Context
-import android.content.Intent
-import androidx.core.content.ContextCompat.startActivity
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.anywherefitness.Api.UserApiBuilder
 import com.example.anywherefitness.model.CreateUser
-import com.example.anywherefitness.model.User
-import com.example.anywherefitness.ui.LoginActivity
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -29,7 +24,6 @@ class RegisterViewModel: ViewModel() {
                     println(response)
                 }
             }
-
         })
     }
 }
